+++ changelog.md
@@ -5,9 +5,11 @@
 - 使用Prettier刷一遍代码
 - 新增自定义属性
 - 设置面板中未设置的选项使用默认值
+- 修改双标签栏的显示效果
 - Use Prettier to format code
 - Add custom attribute
 - The options not set in setting panel uses their default value
+- Change display of the double tab bar
 
 ## v2.2.11